docs_manager/pyproject.toml
[project]
name = "docs-manager"
version = "0.1.0"
description = "Documentation manager for crawling and indexing documentation sites"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "langchain-openai>=0.1.0",
    "langchain-pinecone>=0.2.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "langchain>=0.1.0",
    "crawl4ai>=0.6.0",
    "pinecone>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [] 